integer cors

+++ src/main/java/com/maciejak/ztbdhive/dao/IUserDao.java
@@ -7,13 +7,13 @@ import com.maciejak.ztbdhive.model.User;
 public interface IUserDao {
 	public List<User> getAllUsers();
 
-	public User getUser(Long id);
+	public User getUser(Integer id);
 
 	public User addUser(User user);
 
 	public User editUser(User user);
 
-	public void deleteUser(Long id);
+	public void deleteUser(Integer id);
 	
 	public boolean userExists(User user);
 }
